Split the NEON half-precision tier into NEON_HP and NEON_FHM
NEON_HP.cpp was one translation unit compiled with -march=armv8.2-a+fp16fml,
but its HP-only entry points were dispatched on features.asimdhp alone. Because
the whole TU carried +fp16fml, the compiler was licensed to emit FMLAL/FMLSL
instructions anywhere in it, including into the HP-only functions whose source
has no FMLAL intrinsic. Measured on arm-r8g.xlarge with gcc 12: the HP-only
wrappers compiled from identical source went from 32 AdvSIMD FMLAL/FMLSL
instructions at +fp16fml down to 0 once compiled at +fp16 alone. On a core with
asimdhp but without asimdfhm, the old HP path was therefore a SIGILL.

Tightening the predicate to require both asimdhp and asimdfhm would not have
fixed this: it would have deleted the HP-only fallback for exactly the CPUs
that need it. The fix is two tiers with two translation units, each compiled
only with the license its own kernels need: NEON_HP.cpp now builds at
+fp16, and the new NEON_FHM.cpp carries the FHM-only entry points at
+fp16fml, where they still measure 64 AdvSIMD FMLAL/FMLSL instructions, so the
fast path is intact. Dispatch sites gain a second, independently guarded
branch (asimdhp && asimdfhm) that tries NEON_FHM first and falls back to the
existing asimdhp-only NEON_HP branch.
