+++ cmake/aarch64InstructionFlags.cmake
@@ -9,7 +9,8 @@ CHECK_CXX_COMPILER_FLAG("-march=armv8-a" CXX_ARMV8A)
 CHECK_CXX_COMPILER_FLAG("-march=armv8.2-a+dotprod" CXX_NEON_DOTPROD)
 CHECK_CXX_COMPILER_FLAG("-march=armv8-a+sve" CXX_SVE)
 CHECK_CXX_COMPILER_FLAG("-march=armv9-a+sve2" CXX_SVE2)
-CHECK_CXX_COMPILER_FLAG("-march=armv8.2-a+fp16fml" CXX_NEON_HP)
+CHECK_CXX_COMPILER_FLAG("-march=armv8.2-a+fp16" CXX_NEON_HP)
+CHECK_CXX_COMPILER_FLAG("-march=armv8.2-a+fp16fml" CXX_NEON_FHM)
 CHECK_CXX_COMPILER_FLAG("-march=armv8.2-a+bf16" CXX_NEON_BF16)
 CHECK_CXX_COMPILER_FLAG("-march=armv8.2-a+sve+bf16" CXX_SVE_BF16)
 
@@ -28,6 +29,10 @@ if (CXX_ARMV8A OR CXX_ARMV7_NEON)
 if (CXX_NEON_HP)
   message(STATUS "Using ARMv8.2-a with NEON half-percision extension")
   add_compile_definitions(OPT_NEON_HP)
+endif()
+if (CXX_NEON_FHM)
+  message(STATUS "Using ARMv8.2-a with NEON FHM extension")
+  add_compile_definitions(OPT_NEON_FHM)
 endif()
 if (CXX_NEON_BF16)
   add_compile_definitions(OPT_NEON_BF16)
