+++ src/VecSim/spaces/functions/NEON_FHM.h
@@ -1,0 +1,19 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+ */
+#pragma once
+
+#include "VecSim/spaces/spaces.h"
+
+namespace spaces {
+
+dist_func_t<float> Choose_SQ8_FP16_IP_implementation_NEON_FHM(size_t dim);
+dist_func_t<float> Choose_SQ8_FP16_L2_implementation_NEON_FHM(size_t dim);
+dist_func_t<float> Choose_SQ8_FP16_Cosine_implementation_NEON_FHM(size_t dim);
+
+} // namespace spaces
