+++ tests/unit/test_spaces.cpp
@@ -43,6 +43,7 @@
 #include "VecSim/spaces/functions/NEON.h"
 #include "VecSim/spaces/functions/NEON_DOTPROD.h"
 #include "VecSim/spaces/functions/NEON_HP.h"
+#include "VecSim/spaces/functions/NEON_FHM.h"
 #include "VecSim/spaces/functions/NEON_BF16.h"
 #include "VecSim/spaces/functions/SVE.h"
 #include "VecSim/spaces/functions/SVE_BF16.h"
@@ -3370,7 +3371,7 @@ TEST_P(SQ8_FP16_SpacesOptimizationTest, SQ8_FP16_L2SqrTest) {
         optimization.sve = 0;
     }
 #endif
-#ifdef OPT_NEON_HP
+#ifdef OPT_NEON_FHM
     if (optimization.asimdfhm) {
         unsigned char alignment = 0;
         arch_opt_func = L2_SQ8_FP16_GetDistFunc(dim, &alignment, &optimization);
@@ -3381,6 +3382,8 @@ TEST_P(SQ8_FP16_SpacesOptimizationTest, SQ8_FP16_L2SqrTest) {
         ASSERT_EQ(alignment, 0) << "No alignment NEON_FHM with dim " << dim;
         optimization.asimdfhm = 0;
     }
+#endif
+#ifdef OPT_NEON_HP
     if (optimization.asimdhp) {
         unsigned char alignment = 0;
         arch_opt_func = L2_SQ8_FP16_GetDistFunc(dim, &alignment, &optimization);
@@ -3494,7 +3497,7 @@ TEST_P(SQ8_FP16_SpacesOptimizationTest, SQ8_FP16_InnerProductTest) {
         optimization.sve = 0;
     }
 #endif
-#ifdef OPT_NEON_HP
+#ifdef OPT_NEON_FHM
     if (optimization.asimdfhm) {
         unsigned char alignment = 0;
         arch_opt_func = IP_SQ8_FP16_GetDistFunc(dim, &alignment, &optimization);
@@ -3505,6 +3508,8 @@ TEST_P(SQ8_FP16_SpacesOptimizationTest, SQ8_FP16_InnerProductTest) {
         ASSERT_EQ(alignment, 0) << "No alignment NEON_FHM with dim " << dim;
         optimization.asimdfhm = 0;
     }
+#endif
+#ifdef OPT_NEON_HP
     if (optimization.asimdhp) {
         unsigned char alignment = 0;
         arch_opt_func = IP_SQ8_FP16_GetDistFunc(dim, &alignment, &optimization);
@@ -3618,7 +3623,7 @@ TEST_P(SQ8_FP16_SpacesOptimizationTest, SQ8_FP16_CosineTest) {
         optimization.sve = 0;
     }
 #endif
-#ifdef OPT_NEON_HP
+#ifdef OPT_NEON_FHM
     if (optimization.asimdfhm) {
         unsigned char alignment = 0;
         arch_opt_func = Cosine_SQ8_FP16_GetDistFunc(dim, &alignment, &optimization);
@@ -3629,6 +3634,8 @@ TEST_P(SQ8_FP16_SpacesOptimizationTest, SQ8_FP16_CosineTest) {
         ASSERT_EQ(alignment, 0) << "No alignment NEON_FHM with dim " << dim;
         optimization.asimdfhm = 0;
     }
+#endif
+#ifdef OPT_NEON_HP
     if (optimization.asimdhp) {
         unsigned char alignment = 0;
         arch_opt_func = Cosine_SQ8_FP16_GetDistFunc(dim, &alignment, &optimization);
